Updates query

+++ updateAssenza.php
@@ -9,8 +9,7 @@
 	
 	$mode = $_POST['mode'];
 	$id = $_POST['id'];
-	$date = $_POST['data'];
-	$motivation = $_POST['motivazione'];
+	$motivation = $_POST['motivation'];
 	
 	$query = '';
 	$array = [];
@@ -20,41 +19,35 @@
 			if(!isset($_POST['motivation'])) error("invalid");
 			if($_SESSION['adult'] || check_role('parent')){
 				if(check_role('parent')){
-					/*//prende la matricola dello studente minorenne
+					//prende la matricola dello studente minorenne
 					$sonParentNumb = $SESSION['son'];
 
-					//prende la matricola dello studente proprietario dell'evento
-					$query = "SELECT U_Matricola FROM evento WHERE Id = ?";
+					//prende la matricola dello studente proprietario dell'evento se la matricola è la stessa della precedente
+					$query = "SELECT U_Matricola FROM evento WHERE Id = ? AND U_Matricola = ?";
 					$result = $db->prepare($query);
-    				$result->execute([$id]);
+    				$result->execute([$id, $sonParentNumb]);
     
-					$row = $result->fetch();
-					$eventStudentNumb = $row['U_Matricola'];
-					*/
-					//verifica se la matricola dello studente minorenne e quella dello studente proprietario dell'evento combacino
-					//if($sonParentNumb==$eventStudentNumb){
+					//verifica l'esito della query
+					if($result->rowCount() == 0){
+						error("not_parent");
+					}else{
 						$query = "UPDATE evento SET Motivazione = ?, Stato = 'In attesa' WHERE (Id = ?)";
 						$array = [$motivation, $id];
-					//}else{
-					//	error("not_parent");
-					//}
+					}
 				}else
 					if($_SESSION['adult']){
 						//prende la matricola dello studente proprietario dell'evento
-						/*$query = "SELECT U_Matricola FROM evento WHERE Id = ?";
+						$query = "SELECT U_Matricola FROM evento WHERE Id = ? AND U_Matricola = ?";
 						$result = $db->prepare($query);
-						$result->execute([$id]);
-		
-						$row = $result->fetch();
-						$eventStudentNumb = $row['U_Matricola'];
-						*/
+						$result->execute([$id, $_SESSION['user']]);
+						
 						//verifica se la matricola dello studente in sessione e quella dello studente proprietario dell'evento combacino
-						//if($_SESSION['user']==$eventStudentNumb){
+						if($stm->rowCount() == 0){
+							error("not_owner");
+						}else{
 							$query = "UPDATE evento SET Motivazione = ?, Stato = 'In attesa' WHERE (Id = ?)";
 							$array = [$motivation, $id];
-						//}else{
-						//	error("not_owner");
-						//}
+						}
 					}
 			}else{
 				error("insufficient_permission");
@@ -63,14 +56,6 @@
 		
 		case 'r': //remove
 			if(check_role('admin')){
-				/*//prende la classe di appartenenza dell'admin
-				$query = "SELECT U_Matricola, C_Id FROM appartenere INNER JOIN classi ON (appartenere.C_Id = classe.Id) WHERE U_Matricola = ?";
-				$result = $db->prepare($query);
-				$result->execute([$_SESSION['user']]);
-
-				$row = $result->fetch();
-				$adminClass = $row['C_Id'];
-
 				//prende la matricola dello studente proprietario dell'evento
 				$query = "SELECT U_Matricola FROM evento WHERE (Id = ?)";
 				$result = $db->prepare($query);
@@ -86,14 +71,19 @@
 
 				$row = $result->fetch();
 				$studentClass = $row['C_Id'];
-				*/
+				
+				//prende la classe di appartenenza dell'admin
+				$query = "SELECT U_Matricola, C_Id FROM appartenere INNER JOIN classi ON (appartenere.C_Id = classe.Id) WHERE U_Matricola = ? AND C_Id = ?";
+				$result = $db->prepare($query);
+				$result->execute([$_SESSION['user'], $studentClass]);
+			
 				//verifica che la classe dell'admin in sessione combaci con quella dello studente proprietario dell'evento
-				//if($adminClass == $studentClass){
+				if($stm->rowCount() == 0){
+					error("not_same_class");
+				}else{
 					$query = "DELETE FROM evento WHERE (Id = ?)";
 					$array = [$id];
-				//}else{
-				//	error("not_same_class");
-				//}
+				}
 			}else{
 				error("insufficient_permission");
 			}
@@ -102,14 +92,6 @@
 		case 'a': //approve
 			if(!isset($_POST['motivation'])) error("invalid");
 			if(check_role('admin')){
-				//prende la classe di appartenenza dell'admin
-				/*$query = "SELECT U_Matricola, C_Id FROM appartenere INNER JOIN classi ON (appartenere.C_Id = classe.Id) WHERE U_Matricola = ?";
-				$result = $db->prepare($query);
-				$result->execute([$_SESSION['user']]);
-
-				$row = $result->fetch();
-				$adminClass = $row['C_Id'];
-
 				//prende la matricola dello studente proprietario dell'evento
 				$query = "SELECT U_Matricola FROM evento WHERE (Id = ?)";
 				$result = $db->prepare($query);
@@ -125,13 +107,18 @@
 
 				$row = $result->fetch();
 				$studentClass = $row['C_Id'];
-				*/
-				//if($adminClass == $studentClass){
+				
+				//prende la classe di appartenenza dell'admin
+				$query = "SELECT U_Matricola, C_Id FROM appartenere INNER JOIN classi ON (appartenere.C_Id = classe.Id) WHERE U_Matricola = ? AND C_Id = ?";
+				$result = $db->prepare($query);
+				$result->execute([$_SESSION['user'], $studentClass]);
+
+				if($stm->rowCount() == 0){
+					error("not_same_class");
+				}else{
 					$query = "UPDATE evento SET Stato = 'Accettato' WHERE (Id = ?)";
 					$array = [$id];
-				//}else{
-				//	error("not_same_class");
-				//}
+				}
 			}else{
 				error("insufficient_permission");
 			}
@@ -140,14 +127,6 @@
 		case 'd': //deny
 			if(!isset($_POST['motivation'])) error("invalid");
 			if(check_role('admin')){
-				//prende la classe di appartenenza dell'admin
-				/*$query = "SELECT U_Matricola, C_Id FROM appartenere INNER JOIN classi ON (appartenere.C_Id = classe.Id) WHERE U_Matricola = ?";
-				$result = $db->prepare($query);
-				$result->execute([$_SESSION['user']]);
-
-				$row = $result->fetch();
-				$adminClass = $row['C_Id'];
-
 				//prende la matricola dello studente proprietario dell'evento
 				$query = "SELECT U_Matricola FROM evento WHERE (Id = ?)";
 				$result = $db->prepare($query);
@@ -163,18 +142,24 @@
 
 				$row = $result->fetch();
 				$studentClass = $row['C_Id'];
-				*/
-				//if($adminClass == $studentClass){
+
+				//prende la classe di appartenenza dell'admin
+				$query = "SELECT U_Matricola, C_Id FROM appartenere INNER JOIN classi ON (appartenere.C_Id = classe.Id) WHERE U_Matricola = ? AND C_Id = ?";
+				$result = $db->prepare($query);
+				$result->execute([$_SESSION['user'], $studentClass]);
+				
+				if($stm->rowCount() == 0){
+					error("not_same_class");
+				}else{
 					$query = "UPDATE evento SET Stato = 'Rifiutato' WHERE (Id = ?)";
 					$array = [$id];
-				//}else{
-				//	error("not_same_class");
-				//}
+				}
 			}else{
 				error("insufficient_permission");
 			}
 		break;
 	}
+	
 	var_dump($mode);
 	var_dump($query);
 	var_dump($array);
